fantasychess-server/src/main/java/com/bteam/fantasychess_server/handler/TextWebSocketHandlerExt.java
package com.bteam.fantasychess_server.handler;

import com.bteam.fantasychess_server.service.WebSocketService;
import org.springframework.web.socket.CloseStatus;
import org.springframework.web.socket.TextMessage;
import org.springframework.web.socket.WebSocketSession;
import org.springframework.web.socket.handler.TextWebSocketHandler;

public class TextWebSocketHandlerExt extends TextWebSocketHandler {
    WebSocketService service;

    public TextWebSocketHandlerExt(WebSocketService service) {
        this.service = service;
    }

    @Override
    public void afterConnectionEstablished(WebSocketSession session) throws Exception {
        super.afterConnectionEstablished(session);
        service.registerSession(session);
    }

    @Override
    public void handleTextMessage(WebSocketSession session, TextMessage message) throws Exception {
        super.handleTextMessage(session, message);
        service.handleTextMessage(session, message);
    }

    @Override
    public void afterConnectionClosed(WebSocketSession session, CloseStatus status) throws Exception {
        super.afterConnectionClosed(session, status);
        service.removeSession(session.getId(), status);
    }
}
